+++ fileManager.js
@@ -1,6 +1,6 @@
 import readline from 'readline';
 import { parseUsername } from './src/utils/parseUsername.js';
-import { getCurrentDirectory } from './src/config.js';
+import { getCurrentDirectory } from './src/utils/getCurrentDir.js';
 import { handleCommand } from './src/commandOperations/handleCommand.js';
 
 const readLine = readline.createInterface({
