+++ src/commands/nwd/ls.js
@@ -1,5 +1,5 @@
 import fs from 'fs/promises';
-import { getCurrentDirectory } from '../../config.js';
+import { getCurrentDirectory } from '../../utils/getCurrentDir.js';
 
 export const listDirectoryContent = async () => {
   try {
